+++ runner.sh
@@ -1,2 +1,0 @@
-#!/usr/bin/sh
-poetry run bookman -c test-files/config.py $@